# backend/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "AcademiaConnect API",
                 "description":  "Backend API for AcademiaConnect SIH 2026",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/opportunities/":  {
                                          "get":  {
                                                      "tags":  [
                                                                   "Opportunities"
                                                               ],
                                                      "summary":  "Get Opportunities",
                                                      "operationId":  "get_opportunities_opportunities__get",
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "Successful Response",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {

                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                }
                                                                    }
                                                  },
                                          "post":  {
                                                       "tags":  [
                                                                    "Opportunities"
                                                                ],
                                                       "summary":  "Create Opportunity",
                                                       "operationId":  "create_opportunity_opportunities__post",
                                                       "requestBody":  {
                                                                           "content":  {
                                                                                           "application/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/app__api__routes__opportunities__OpportunityCreate"
                                                                                                                               }
                                                                                                                }
                                                                                       },
                                                                           "required":  true
                                                                       },
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "Successful Response",
                                                                                     "content":  {
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {

                                                                                                                                         }
                                                                                                                          }
                                                                                                 }
                                                                                 },
                                                                         "422":  {
                                                                                     "description":  "Validation Error",
                                                                                     "content":  {
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                         }
                                                                                                                          }
                                                                                                 }
                                                                                 }
                                                                     }
                                                   }
                                      },
                  "/opportunities/{opportunity_id}":  {
                                                          "get":  {
                                                                      "tags":  [
                                                                                   "Opportunities"
                                                                               ],
                                                                      "summary":  "Get Opportunity",
                                                                      "operationId":  "get_opportunity_opportunities__opportunity_id__get",
                                                                      "parameters":  [
                                                                                         {
                                                                                             "name":  "opportunity_id",
                                                                                             "in":  "path",
                                                                                             "required":  true,
                                                                                             "schema":  {
                                                                                                            "type":  "integer",
                                                                                                            "title":  "Opportunity Id"
                                                                                                        }
                                                                                         }
                                                                                     ],
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "Successful Response",
                                                                                                    "content":  {
                                                                                                                    "application/json":  {
                                                                                                                                             "schema":  {

                                                                                                                                                        }
                                                                                                                                         }
                                                                                                                }
                                                                                                },
                                                                                        "422":  {
                                                                                                    "description":  "Validation Error",
                                                                                                    "content":  {
                                                                                                                    "application/json":  {
                                                                                                                                             "schema":  {
                                                                                                                                                            "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                        }
                                                                                                                                         }
                                                                                                                }
                                                                                                }
                                                                                    }
                                                                  },
                                                          "delete":  {
                                                                         "tags":  [
                                                                                      "Opportunities"
                                                                                  ],
                                                                         "summary":  "Delete Opportunity",
                                                                         "operationId":  "delete_opportunity_opportunities__opportunity_id__delete",
                                                                         "parameters":  [
                                                                                            {
                                                                                                "name":  "opportunity_id",
                                                                                                "in":  "path",
                                                                                                "required":  true,
                                                                                                "schema":  {
                                                                                                               "type":  "integer",
                                                                                                               "title":  "Opportunity Id"
                                                                                                           }
                                                                                            }
                                                                                        ],
                                                                         "responses":  {
                                                                                           "200":  {
                                                                                                       "description":  "Successful Response",
                                                                                                       "content":  {
                                                                                                                       "application/json":  {
                                                                                                                                                "schema":  {

                                                                                                                                                           }
                                                                                                                                            }
                                                                                                                   }
                                                                                                   },
                                                                                           "422":  {
                                                                                                       "description":  "Validation Error",
                                                                                                       "content":  {
                                                                                                                       "application/json":  {
                                                                                                                                                "schema":  {
                                                                                                                                                               "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                           }
                                                                                                                                            }
                                                                                                                   }
                                                                                                   }
                                                                                       }
                                                                     }
                                                      },
...
                  "/company/opportunities/":  {
                                                  "post":  {
                                                               "tags":  [
                                                                            "Company Opportunities"
                                                                        ],
                                                               "summary":  "Create Opportunity",
                                                               "operationId":  "create_opportunity_company_opportunities__post",
                                                               "requestBody":  {
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/app__api__routes__company_opportunities__OpportunityCreate"
                                                                                                                                       }
                                                                                                                        }
                                                                                               },
                                                                                   "required":  true
                                                                               },
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "Successful Response",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {

                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "422":  {
                                                                                             "description":  "Validation Error",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         }
                                                                             }
                                                           }
                                              },
                  "/company/opportunities/{user_id}":  {
                                                           "get":  {
                                                                       "tags":  [
                                                                                    "Company Opportunities"
                                                                                ],
                                                                       "summary":  "Get Company Opportunities",
                                                                       "operationId":  "get_company_opportunities_company_opportunities__user_id__get",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "user_id",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "integer",
                                                                                                             "title":  "User Id"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "Successful Response",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {

                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 },
                                                                                         "422":  {
                                                                                                     "description":  "Validation Error",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 }
                                                                                     }
                                                                   }
                                                       },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           },
...
                                   }
                   }
}